requirements.txt
streamlit==1.39.0
opencv-python==4.7.0
torch==2.0.0+cpu
torchvision==0.15.1+cpu
numpy==1.23.5
Pillow==10.0.0
